refactor: extract ChatMessage as public component
