+++ web/src/routes/share/index.jsx
@@ -4,9 +4,9 @@ import { useContext } from "react";
 import { useLoaderData, redirect } from "react-router-dom";
 
 import ChatLog from "components/ChatLog";
+import ChatMessage from "components/ChatMessage";
 import { ThemeContext } from "contexts/theme";
 
-import ChatMessage from "../conversation/ChatMessage";
 
 export async function loader({ params }) {
     const resp = await fetch(`/api/shares/${params.shareId}`, {});
